refactor: quit unnused diles, console.logs and update imports
