--- backend/client-gateaway/src/courses/courses.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { CoursesService } from './courses.service';
-
-describe('CoursesService', () => {
-  let service: CoursesService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [CoursesService],
-    }).compile();
-
-    service = module.get<CoursesService>(CoursesService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
