--- backend/client-gateaway/src/courses/dto/create-course.dto.ts
@@ -1,1 +1,0 @@
-export class CreateCourseDto {}
