--- backend/client-gateaway/src/courses/dto/update-course.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateCourseDto } from './create-course.dto';
-
-export class UpdateCourseDto extends PartialType(CreateCourseDto) {}
