--- backend/client-gateaway/src/courses/entities/course.entity.ts
@@ -1,1 +1,0 @@
-export class Course {}
