--- backend/client-gateaway/src/inscriptions/dto/create-inscription.dto.ts
@@ -1,1 +1,0 @@
-export class CreateInscriptionDto {}
