--- backend/client-gateaway/src/inscriptions/entities/inscription.entity.ts
@@ -1,1 +1,0 @@
-export class Inscription {}
