--- backend/client-gateaway/src/users/entities/user.entity.ts
@@ -1,1 +1,0 @@
-export class User {}
